Move AddProjects back to components folder
